Install JsThemis without sudo
Using "sudo make jsthemis_install" will result in system Node.js being
used, not the one installed for testing. Run JsThemis installation as
a regular user. Ditto for WasmThemis (though it should not be affected).

The reason it needs to be run with sudo sometimes is that some previous
installer builds Themis as root so the build directory ends up being
owned by root and npm cannot move its stuff there. Apply a quickfix for
that, but if we do it properly, we should not be building stuff as root
in the first place.

I'm really seriously frustrated with this changeset so I don't have
mental capacity to debug this tangle at the moment. I'll leave a FIXME
there and hope to come back at it later.