Improve git ref hacks in ObjCThemis (#610)
It turns out that GitHub uses a temporary ref for building pull requests
that contains the pull request head merged with the base branch.
Unfortunately, CocoaPods cannot fetch that commit because it does not
fetch all git refs, only branches. This means that all pull requests
will fail as CocoaPods is not able to fetch them.

Instead of using GITHUB_REPOSITORY and GITHUB_SHA, use tweaked versions
of them when building pull requests. They will tell CocoaPods and
Carthage to fetch the branch submitted in pull request, not the result
of merge of that branch with the base branch.

It's kind of stupid but I have no other ideas on how to test builds that
depend on git refs in the repo itself. Blame CocoaPods, IDK vOv

Also, we should trigger this workflow when "Cartfile" is changed, not
"Carthage" (which is the build directory).