Wipe Secure Cell key copy on deallocation
Secure Cell currently makes a copy of the master key it was given. Since
Objective-C is fairly low level and exposes the "key" property, it makes
sense to wipe the sensitive key from memory when it can no longer be
used by Secure Cell.

Do so by changing the underlying property type to NSMutableData and
calling the wiping code in "dealloc" of TSCell.

While we're here, improve API docs of the basic Secure Cell class that
actually stores the key.

Note that in Objective-C the "key" property returns NSData which
respects retain-release mechanics so the users get a reference to
the same data object as used by Secure Cell. However, in Swift NSData
is bridged into Data type which has value semantics and effectively
copies our copy of the key. We obviously cannot and should not wipe
the copies we are not aware of, so this is a best effort approach.