Remove SCELL_COMPAT define and ignore NO_SCELL_COMPAT variable
It does not really make much sense for the users to not enable the
compatibility code path that allows to decrypt historical data.
In fact, not a single build that we distribute disables this code
path.

Remove all SCELL_COMPAT ifdefs and ignore NO_SCELL_COMPAT setting,
making the compatibility always enabled.

This has a side effect of slightly slowing down decryption of really
corrupted messages since we will be always trying the fallback path.
However, this is not that significant loss as originally envisioned,
and normal usage of Themis should not be affected.