Update to actions/checkout@v2
Some of our build jobs *cough* Android *cough* occasionally need to be
restarted. One issue that I've noticed is that for restarted jobs the
checkout task tries to fetch a different commit (PR branch HEAD, not
the result of merge with the base branch). Since v1 is unlikely to be
maintained, let's upgrade to v2 and see if that issue persists.