Secure Cell passphrase API: ObjCThemis (#609)
* Passphrase API support in ObjCThemis

Add passphrase API support for Objective-C and Swift. The new private
subclass TSCellSealWithPassphrase implements encryption and decryption
with passphrases while TSCellSeal provides initializers.

This is not how you do class clusters in Objective-C, but unfortunately
this is the only way to get a sane Swift API without actually writing
Swift code. We cannot include Swift code in ObjCThemis because Swift
lacks ABI stability and Xcode projects require a particular version of
Swift to be set.

Swift 5 supports stable ABI but we cannot do much about this while we
have to support Swift 4 (which we will probably do until it is available
in Xcode).

Current implementation is usable as is but it is currently impossible
for users to inherit from TSCellSealWithPassphrase. This should not be
necessary, it seems, but there are legitimate cases when this might be
requested (i.e., common base64-encoded encryption). We'll deal with that
later based on user feedback. It is possible to export the private
subclass to allow subclassing it, but initialization via TSCellSeal is
still much more convenient.

* Remove the initWithPassphrase:usingEncoding: method

Since it's not possible to bridge this method into Swift, let's drop it
completely. Replace the test with the same as in Swift which checks that
we are able to decrypt data with UTF-16BE passphrase.