Correct data corruption tests (#613)
Replace "256 - byte" with "~byte" as an example of data corruption.
I don't know what the person writing that thought, but if the byte
happens to be equal exactly 128 then 256 - 128 = 128, nothing gets
corrupted, and the tests fail. They do occasionally fail, when the
stars are just right, this was attributed to some issue in the token
parsers of Secure Cell which ignored some of the bytes. Now all that
code has been updated to be more strict so *any* changes in the data
should be reliably detected and result in a failure.