New construction API and passphrase support
Add new static factory methods "withKey()" to all Secure Cell classes,
with SecureCellSeal getting additional "withPasshrase()" to support
passphrase API.

The new API allows to disambiguate between master keys and passphrases
better then the current new constructor.

Passphrase API is implemented in a subclass of SecureCellSeal so that
all typing tests still work in JavaScript. The class itself is not
exported, the only way to construct it is through the withPassword()
method. That might change in the future.

There's quite a lot of copy-pasta involved in the implementation but for
now we did not find any convenient way to avoid it. Note that like
ObjCThemis, the "passphrase bytes" are kept in the "masterKey" property
of Secure Cell object. Themis Core does not really care, as long as they
are in an Uint8Array.

Also note TextEncoder usage to convert strings into UTF-8. JavaScript
strings are normally in UTF-16. TextEncoder is supported by modern Web
browsers but only for UTF-8. Node.js allows other encodings, but we're
not interested in them right now. Being limited with UTF-8 is the reason
for why we do not provide an out-of-the-box support for other encodings
like some other Themis wrappers do.