Update existing Secure Cell tests
Update the tests to use the new construction syntax.

As for

> It's a really nice idea to accept strings as 'master keys'

this is not entirely true. This comment has been written at the time
when we believed that Secure Cell can be safely used with passphrases
via its master key API. Remove these comments, master key API is not
going to support strings anymore.