Update to actions/checkout@v2 (#615)
* Update to actions/checkout@v2

Some of our build jobs *cough* Android *cough* occasionally need to be
restarted. One issue that I've noticed is that for restarted jobs the
checkout task tries to fetch a different commit (PR branch HEAD, not
the result of merge with the base branch). Since v1 is unlikely to be
maintained, let's upgrade to v2 and see if that issue persists.

* Install git in "code-style" workflow

Since we're using a custom Docker image here, we need to have actual
git execuable available for actions/checkout@v2 to work properly with
submodules. Version 1 was able to use GitHub REST API but the new
version seems to have issues with that. Provide it with real git.

* Switch to Ubuntu 19.10 for code style checks

actions/checkout@v2 with submodules requires git 2.18+. Ubuntu 18.04
has only 2.17 in its repos, so upgrade to currently supported 19.10.
Hopefully, we'll switch to the next LTS 20.04 as soon as it's out.

With Ubuntu 19.10 system Node.js installation no longer depends on
outdated OpenSSL 1.0.2 so we can install the regular "libssl-dev".