Test all matrix configurations (#617)
Add "fail-fast: false" to matrix testing strategies so that they
continue testing other variant instead of auto-canceling them.
For example, if only Node.js v8 fails, we don't want the Node.js v12
job to be canceled. This might be different in future cases, but
everywhere we use "strategy: matrix" now, we'd like all items to be
tested regardless of their siblings' failures.