Move and simplify JsThemis installation targets
First of all, move them into the dedicated makefile from the main one.
Introduce a proper file target that will build the package in the build
directory (not simply "./build"). This is more friendly to systems where
the source directory is mounted as read-only, with a separate writable
build directory.

Unfortunately, "npm pack" always spits out its results into the current
directory. We have to cd into the build directory before calling it.

Also, since we don't really need to have multiple versions in the build
directory, we can rename the result into just "jsthemis.tgz" and avoid
doing all that awk magic to extract the version.

Note the use of "abspath". We need to ensure that we refer to correct
files, and for "npm install" to install the archive without treating
"build/jsthemis.tgz" as a scoped package name.