Convenience "jsthemis" target
Add a new target "jsthemis" that can be used to build JsThemis tarball
in the build directory. Finally we have something to provide as
instructions on how to build JsThemis from source.

Make sure "npm pack" does not output not-so-correct filename to stdout.
If only there was a way to make it put its output into a different
directory...