Move test directory into JsThemis source tree
npm seems to be really insinstent on that we either publish the thing we
want to test on npm, or use relative paths.

An alternative to this is using "npm link" to link the current directory
into local node_modules, but that 1) modifies global npm state, 2) does
not work reliably if run twice as for some reason it removes package
dependencies on the second run.

So I gave up. Move the test directory from top-level "tests" directly
into JsThemis source tree, just like we do with WasmThemis (for the same
reason).