Secure Cell passphrase API: WasmThemis (#616)
* New construction API and passphrase support

Add new static factory methods "withKey()" to all Secure Cell classes,
with SecureCellSeal getting additional "withPasshrase()" to support
passphrase API.

The new API allows to disambiguate between master keys and passphrases
better then the current new constructor.

Passphrase API is implemented in a subclass of SecureCellSeal so that
all typing tests still work in JavaScript. The class itself is not
exported, the only way to construct it is through the withPassword()
method. That might change in the future.

There's quite a lot of copy-pasta involved in the implementation but for
now we did not find any convenient way to avoid it. Note that like
ObjCThemis, the "passphrase bytes" are kept in the "masterKey" property
of Secure Cell object. Themis Core does not really care, as long as they
are in an Uint8Array.

Also note TextEncoder usage to convert strings into UTF-8. JavaScript
strings are normally in UTF-16. TextEncoder is supported by modern Web
browsers but only for UTF-8. Node.js allows other encodings, but we're
not interested in them right now. Being limited with UTF-8 is the reason
for why we do not provide an out-of-the-box support for other encodings
like some other Themis wrappers do.

* Update existing Secure Cell tests

Update the tests to use the new construction syntax.

As for

> It's a really nice idea to accept strings as 'master keys'

this is not entirely true. This comment has been written at the time
when we believed that Secure Cell can be safely used with passphrases
via its master key API. Remove these comments, master key API is not
going to support strings anymore.

* Test passphrase API

Add tests for passphrase API, taking into account its peculiarities and
incompatibility with the master key API.

* Expect passphrase API timings

Due to KDF usage, passphrase API is intentionally slow. However, Mocha
testing framework has too low default threshold for slow tests. Bump it
up a bit so that we're not risking to fail the test.

Also, ensure that passphrase API *is* slow enough, just in case it gets
swapped with master key API, or some technical progress will make
WebAssembly to run really fast.

Currently, bare-metal benchmarks of Themis Core suggest that default
iteration count for PBKDF2 results in 125 ms median time for encryption
and decryption. WebAssembly shows about 400-500 ms median. Well, not bad
but not amazing either.

* Changelog entry

* Polyfill TextEncoder

O, JavaScript, language of a thousand incompatibilities! You managed
to bring Lisp to the Web, with all its merits and vices!

Long story short, Node.js v10 and some outdated browsers do not support
TextEncoder API. Add a polyfill to provide TextEncoder API on those
platforms.

We don't want to write yet another UTF-8 encoder ourselves to let's use
a library. This is our first external dependency in WasmThemis. Here are
some obligatory links:

  - https://github.com/anonyco/FastestSmallestTextEncoderDecoder
  - https://developer.mozilla.org/en-US/docs/Web/API/TextEncoder

The library does not have any dependencies itself, is licensed under
the Unlicense (effectively public domain) which is compatible with our
Apache 2.0, has seen some testing, and is quite minimal so we can
maintain it ourselves if necessary. However, I'd consider adding
explicit "npm audit" steps to build jobs from now on...

* Use new syntax in integration tools

* Integration tool for passphrase API

Add a tool to check integration of passphrase API with other platforms.

* Update precaution phrasing in changelog

Give more specific instructions on when to use master key API vs
passphrase API. Also, avoid phrasing about master key API being not
secure with passphrases. It *is* secure, given the passphrase of
sufficient length.

Update other languages to follow the style as well.

Co-Authored-By: vixentael <vixentael@users.noreply.github.com>