Simplify JsThemis packaging (#618)
* Improve JsThemis packaging

Add package-lock.json file into the repo. It is recommended to keep this
file committed into VCS. GitHub analyzer will scan it to see if there
are known security vulnerabilities in our dependencies.

Update the package.json file, add some missing metadata fields that we
fill in for WasmThemis. Also, add modern "mocha" as dev dependency.

Add an .npmignore file to prevent internal files from being packed into
the resulting distribution package.

* Move and simplify JsThemis installation targets

First of all, move them into the dedicated makefile from the main one.
Introduce a proper file target that will build the package in the build
directory (not simply "./build"). This is more friendly to systems where
the source directory is mounted as read-only, with a separate writable
build directory.

Unfortunately, "npm pack" always spits out its results into the current
directory. We have to cd into the build directory before calling it.

Also, since we don't really need to have multiple versions in the build
directory, we can rename the result into just "jsthemis.tgz" and avoid
doing all that awk magic to extract the version.

Note the use of "abspath". We need to ensure that we refer to correct
files, and for "npm install" to install the archive without treating
"build/jsthemis.tgz" as a scoped package name.

* Simplify JsThemis unit test launch

Add an npm target "test" to run "mocha" for our tests. This makes is
possible to develop and test JsThemis directly from its source directory
(like WasmThemis) and also packages the tests so that can be used by
other developers to verify our package after installation.

These changes make it possible to remove the preparation step from unit
test runs of JsThemis. We can now simply do "make test_js", without
installing it into the source root and running "make prepare_tests_all".

* Convenience "jsthemis" target

Add a new target "jsthemis" that can be used to build JsThemis tarball
in the build directory. Finally we have something to provide as
instructions on how to build JsThemis from source.

Make sure "npm pack" does not output not-so-correct filename to stdout.
If only there was a way to make it put its output into a different
directory...

* Move test directory into JsThemis source tree

npm seems to be really insinstent on that we either publish the thing we
want to test on npm, or use relative paths.

An alternative to this is using "npm link" to link the current directory
into local node_modules, but that 1) modifies global npm state, 2) does
not work reliably if run twice as for some reason it removes package
dependencies on the second run.

So I gave up. Move the test directory from top-level "tests" directly
into JsThemis source tree, just like we do with WasmThemis (for the same
reason).