Passphrase API for Seal mode
Introduce a "withPassphrase" static factory method which constructs a
subclass of SecureCellSeal that uses passphrase API instead of regular
symmetric keys.

This is exactly the same API WasmThemis has. Tests are also reused from
WasmThemis.

Note that SecureCellSealWithPassphrase is a private subclass so it's
not exported and it's not possible to construct it with "new" or inherit
from it in a normal way.

Node.js supports a bit more encodings for it's Buffer class but we still
decode string passphrases as only UTF-8. Applicatons should use Buffer
manually if they need to encode the passphrase in some other encoding.