Make "associated context" non-optional argument
Initially I have envisioned encryption API like this, emulating
function overload missing in Go:

    func (sc *SecureCellSeal) Encrypt(message []byte, context ...[]byte) ([]byte, error)

which allows the user to omit the context argument entirely:

    encrypted, err := cell.Encrypt(message)

and add it only when necessary:

    encrypted, err := cell.Encrypt(message, context)

instead of having to spell it out:

    encrypted, err := cell.Encrypt(message, nil)

However, it turned out to be too bothersome to implement, maintain,
test, and it's not really consistent with AEAD API of "crypto/cipher".
Furthermore, it tends to interact badly with IDE autocompletion which
does not make "context ...[]byte" argument obvious.

Therefore we make the argument (syntactically) non-optional. However,
the users are able to pass "nil" to omit it when they don't need the
associated context.