Passphrase API for Seal mode
Introduce a new constructor "SealWithPassphrase" which returns a
different instance of Secure Cell that works with passphrases instead of
symmetric keys, using appropriate Themis Core API.

Note that we do not convert the string encoding here. Go's "string" type
is expected to contain UTF-8 data (consistent with what other wrappers
use). However, it can contain arbitrary byte sequences too. We do not
interpret them in any way (e.g., as []rune or whatever) and just use
the bytes as is.