Check "./bindgen.sh" output on CI
Set up a GitHub Actions step to verify that the output of "./bindgen.sh"
script does not change and "src/lib.rs" does not need an update. This
will break the build once changes are necessary. Typically this means
that some new API is exported by Themis or there are some changes in
Bindgen's behavior. Either way we would like to be notified.