Package "./configure" for Rust's libthemis-src
Commit f79cd235 ("configure" script for platform detection, 2020-03-27)
has added a script which is now essential for building Themis with Make.
However, it neglected to add this script to source code vendored with
"libthemis-src" crate of RustThemis. Without that script libthemis-src
will not be able to build Themis Core. Add a symlink to it so that this
script is packaged during publishing.

Package VERSION file too because without it the build will try reading
it anyway and print annoying warnings:

    cat: VERSION: No such file or directory

This file was added in Themis 0.12, 7f181597 (API-versioned library
names, 2019-04-16), but I guess we just lived with those warnings.

It is an open question why CircleCI and GitHub Actions do not see this
issue. On my machine it fails, but on CI the tests succeed:

    running 3 tests
    test tests::build_and_install ... cat: VERSION: No such file or directory
    tput: No value for $TERM and no -T specified
    /bin/bash: ./configure: No such file or directory
    ok
    test tests::build_and_install_to_OUT_DIR ... cat: VERSION: No such file or directory
    tput: No value for $TERM and no -T specified
    /bin/bash: ./configure: No such file or directory
    ok
    test tests::pkg_config_setting ... cat: VERSION: No such file or directory
    tput: No value for $TERM and no -T specified
    /bin/bash: ./configure: No such file or directory
    ok

    test result: ok. 3 passed; 0 failed; 0 ignored; 0 measured; 0 filtered out