Add passphrase API of Secure Cell
Introduce new constructor "SecureCell::with_passphrase" which accepts
textual passphrases and returns a Secure Cell that handles them.

Currently only Seal mode supports passphrases so that's the only mode
that users are allowed to select.

The passphrase is accepted as AsRef<[u8]> to transparently allow use of
raw byte slices and other types, not limiting to AsRef<str> which could
be more appropriate (and more restrictive).

This commit does not update existing master key API for correct use of
symmetric keys and does not update API documentation as well. This will
be done separately.