Drop "bindgen" from libthemis-sys dependencies (#626)
* Drop "bindgen" from libthemis-sys dependencies

Initially it was used to dynamically generate Rust bindings to C code.
I have thought that this is the most idiomatic way to it. It is also
more robust as Bindgen will be working with the actual Themis version
installed on the user system.

However, Bindgen has a lot of transitive dependencies, they take quite a
while to compile, and they need to be compiled by our users too. This is
a constant source of irritation for developers. Users are not really
happy with that as well.

While we try our best to maintain backwards compatibility when it comes
to symbols and dependencies, we generally do not bother with forward
compatibility. And we effectively pin wrappers to the core library.
That is, RustThemis 0.N is guaranteed to work with Themis Core 0.N,
it is also guaranteed to work with Themis Core 0.N+1, and will most
likely work with Themis Core 0.N+2 (unless there is deprecated API
removed). However, the converse is not the case. RustThemis 0.N+1 may
work with Themis 0.N if no new features were added, but otherwise it
might fail to compile.

Thus, RustThemis is effectivelly pinned to Themis Core version, meaning
that their versions must match. If that's the case, we do not need to
generate FFI bindings on the user machine because we know which version
we are shipping with and what symbols are available there. Thankfully,
we also write portable C code so Themis API does not depend on the
target architecture -- which allows us to ship the bindings verbatim.
Otherwise we'd have no choice other that generate them on the fly.

With that in mind, start generating "src/lib.rs" manually with the help
of "bindgen.sh" script. The developers are expected to run it on their
machine to regenerate bindings somewhere before release. The users now
do not need to generate these bindings on their machine, meaning that
we can drop "bindgen" crate dependency and all supporting code.

In the end, this gives *significant* improvements in transitive
dependency count (down to 3, from 43) as well as *ENORMOUS*
improvements in build times:

           Before  After
    Linux   28.43   2.68  Debug
    macOS   52.80   8.47

    Linux   52.08   5.35  Release
    macOS  127.21   8.71

CI builds were a source of pain since they build RustThemis in both
debug and release versions, possibly several times, and in the worst
case they are running with only one virtual CPU so the absolute time
to run the test suite might be several minutes.

* Check "./bindgen.sh" output on CI

Set up a GitHub Actions step to verify that the output of "./bindgen.sh"
script does not change and "src/lib.rs" does not need an update. This
will break the build once changes are necessary. Typically this means
that some new API is exported by Themis or there are some changes in
Bindgen's behavior. Either way we would like to be notified.