Move Android build.gradle file from root
Introduce a subproject ":android" for AndroidThemis builds. We can't
continue using top-level build.gradle for everything if we are going
to use Gradle for Desktop Java too.

This changes the invocation strings for targets which now need to be
qualified (by prefixing ":android:...").

Leave the top-level file with some common definitions. This allows to
not write them in each subproject and to use "./gradlew" right from
the repository root.