Enable Gradle's "configure on demand"
By default Gradle will configure all subprojects of a top-level project.
For us that means that we'll force AndroidThemis configuration even if
the user is trying to build JavaThemis for desktop. That will require
a properly installed and configured Android SDK, completely unnecessary.

Thankfully, Gradle also have an experimental "configure on demand"
mode which will configure only the projects actually requested and
needed for the build. Enable this feature. It is 'experimental' but
this means that it might not work for some projects, but Gradle team
plans to make this mode the default one. It works for us, so it's okay.