Run JUnit tests during Gradle builds
The "build" target already runs tests if they are set up. Configure
JUnit test compilation and runner.

Note that JavaThemis requires JNI library to be loaded. By default it is
installed into a location that is not present in the Java library search
path. We need to explicitly add "/usr/local/lib" to "java.library.path"
property. Also note that Gradle will run JUnit stuff in a separate JVM
so we need to pass the system property to that JVM from the main one.